deprecate .bash_enter and .bash_leave, run scripts from $HOME
It was pointed out that depending on and trusting executable pieces
of shell script sprinkled throughout your filesystem is possibly not
the most secure system, however convenient it may be.

To avoid problems with malicious users, or careless admins or package
maintainers, the scripts now run from $HOME/.smartcd

If a file was previously

    /foo/bar/baz/.bash_enter

it is now

    ~/.smartcd/foo/bar/baz/bash_enter

Note both the ~/.smartcd directory, and the removal of the leading
dot on the filename.

If legacy scripts are found, instructions for migrating them to this
new format will be displayed.

In addition, if stash, unstash, or autostash are run interactively,
they too will print instructions on creating these files.