work around issues with zsh's autonamedirs feature
zsh has a feature which detects if a variable is set to a fully
qualified path, and assumes that the variable name is an alias for the
directory.  a bit of an odd feature, and easy to trigger by accident.
so we need to be more careful about exposing variables outside of
smartcd which may contain a path.  localize _apop_return and
_ashift_return from _smartcd to avoid frequently naming directories
"_apop_return"