run smartcd once, on load
This is an incompatible change in behavior, but arguably one that is
more in-line with what a user would expect.  When the shell is started,
all bash_enter scripts starting from the filesystem root are run, up to
and including the cwd.

One thing this means is that if you open a new shell, for example, a
terminal window tab, which starts you in a directory that has such a
script, it will run automatically.

The caveat to this is if you load all 3 libraries (smartcd, varstash,
and arrays), smartcd will automatically run once, whether you asked it
to or not.  If you don't want that behavior, set SMARTCD_NOINITIAL=1
before loading the libraries.