play nicely with zsh's chpwd
The chpwd hook in zsh fires whenever the working directory is changed,
which smartcd does several times during operation.  So, if the user has
varstash loaded, use it to hide this hook while traveling temporarily.

We could also possibly use chpwd for actually implementing smartcd, but
this is not fully thought out yet, so in the meantime just don't cause
pain