add setup_smartcd_exit_hook to trap EXIT
Running this function will cause smartcd to run your bash_leave
scripts when your shell exits, starting with your cwd all the way
down to /

This is the inverse of the feature which runs bash_enter scripts on
smartcd load starting with /, added in commit 251f53f, and which
happens by default.

This feature could be useful if you are doing things such as starting
and stopping processes based on your scripts and you want to ensure
that things are properly shut down when you log out.