UI overhaul
This changes the user interface in a mostly backwards-compatible
fashion.  All user actions are now available through the "smartcd"
function directly, as sub-actions:

    smartcd cd /
    smartcd edit enter
    smartcd show leave
    smartcd filename enter
    smartcd template install foo
    smartcd setup prompt-hook

This of course means that "alias cd=smartcd" is no longer valid.  Since
this is a very common (and indeed, was recommended) configuration, it
will attempt to recognize this usage and notify the user that they
should update their shell config.

To update your configuration, you should change any instances of:

    alias cd=smartcd
    alias pushd=smartpushd
    alias popd=smartpopd

to

    smartcd setup cd
    smartcd setup pushd
    smartcd setup popd

You are still able to alias these commands themselves, for your
convenience:

    alias be="smartcd edit enter"
    alias bl="smartcd edit leave"