allow 'autostash alias foo=bar' as a one-liner
Previously to create a temporary alias you'd have to do something like
this:

    autostash foo
    alias foo=bar

While this worked, it was weird and felt second class to variable
assignment.  The new syntax is a bit of a hack and technically it
prevents you from stashing a variable named "alias" if you're also going
to stash "foo=bar", so don't do that :-)

Function stashing and declaration remains second-class, but one-lining a
function declaration is kind of nasty anyway, so don't do that either.
To make a temporary function, you can use this:

    autostash myfunc
    myfunc() {
        do stuff
    }