move __PATH__ substitution from template install time to runtime
You could previously put the string __PATH__ in a template, which would
be replaced with the corresponding directory name when the template is
installed, to easily share templates between multiple locations without
modification.

However, it is also useful to be able to do this when creating them
yourself, to avoid having to copy&paste or type out the directory
without any mistakes.  This also removes any need to use relative paths,
which could cause issues when you travel to subdirectories.  An example:

    autostash PATH=__PATH__/bin:$PATH

It's probably worth mentioning for posterity that to accomplish this,
script files are no longer "sourced" in-place, but rather slurped up,
possibly substituted, and then eval-d.  This /should/ not introduce any
differences or regressions, but one can never be too sure.