migrate to a new directory structure under .smartcd
This moves scripts and templates into separate subdirectories.  Having
the scripts directly under .smartcd was a mistake and this corrects it.
The migration should happen automatically.

If anything unexpected with the migration happens, your old files should
be available under ~/.smartcd/archive.

The new structure looks like this

    .smartcd/scripts/dir/bash_enter
    .smartcd/templates/tmplt_name

You should be using "smartcd edit" and/or "smartcd filename" to access
these files, which will continue to work.  If you're constructing the
filenames yourself, you'll need to update that logic.

One of the reasons for this change is that it allows other pieces of
metadata to be stored, one of which will be inode data to be able to
detect if the directory a script was created for has been replaced.