validate inode for directory has not changed
If a directory for which enter and/or leave scripts are configured has
been deleted or renamed, the scripts will not be removed automatically.
This can lead to unexpected and possibly undesirable behavior if you
then create a new directory of the same name, since the old scripts will
start running again.

We now record the inode of the directory each script was run for, and if
a change is detected, a warning will be displayed and the scripts will
not be executed.  The user can remove or edit the scripts and wipe out
the inode file to clear the warning.

If you don't want this behavior, set SMARTCD_NO_INODE=1 in your
configuration or select that option in `smartcd config`.