several ways to re-run the enter scripts of the cwd
Sometimes you may want to "leave" and "enter" the current directory
without actually having to do it.  This can be useful if you have edited
the scripts and you want them to run, or if they perform some action
that you want done for any reason.

This can now be done in two different ways:

    smartcd reenter
    cd .