move library files to .smartcd/lib/core
This moves the bash_smartcd, bash_varstash, and bash_arrays files from
being directly in the user's home directory to being under .smartcd.
Additionally, the bash_ prefix was dropped, as this is not a bash-only
library.  This follows 48cb29c37 as part of a several-step process to
re-structure things to be more flexible.