new feature: helpers
This introduces a powerful new element to smartcd, the concept of
"helpers".  These are essentially sharable templates, designed to be
written and distributed by anyone who wants to do so.  This could be the
author of another piece of software, providing easily used smartcd
hooks, or just someone with a good idea.

Eventually this feature will have built-in install and update support,
with access to a centralized repository of these.

For now, the main action is 'run', and is performed like:

    smartcd helper run <name> [args...]

This is most likely useful when added to an enter script.  The syntax
and arguments which can be passed to one of these is entirely up to its
author, and their respective documentation should be viewed for further
information.