allow system-wide scripts
This builds on 8696ccc, which allowed for system-wide installation of
smartcd.  Although it was possible to install it system-wide, it was not
possible to have scripts that were available for all users, without
the unsafe "legacy mode"

The scripts should be located in /etc/smartcd/scripts, unless
$SMARTCD_SYSTEM_SCRIPTS is set, in which case that directory will be
used instead.  Remember that smartcd is not a security tool, and that
users can could change both the value of this variable, or the smartcd
code itself.

Things that are not done for this, but will probably be addressed in the
future are:

  * Documentation of this functionality

  * A user interface for editing the system scripts from smartcd

  * Handling of inode files for system-wide scripts