more DWIMmy behavior for double-stashing
This is a change in the case where you stash-assign the same variable
twice.  Previously, if you did this, you would get an error:

    $ stash foo=bar
    $ stash foo=baz
    You have already stashed foo, please specify "-f" if you want to
    overwrite another stashed value

This error is suppressed if and only if subsequent stashes are assigning
a value.  When you do this, the new value is assigned but the stashed
value is not replaced.

The primary reason for this is that you may find yourself wanting to
touch the same variable several times in short succession, and the
"right thing" is to allow this to happen but then to restore to the
original value upon unstash.