new `leave-hook` action which can be used to create transient leave scripts
This is called like this:

    smartcd leave-hook 'shell code'

The shell code will be stored in a temporary function which will be
executed on leave from the directory in which it was defined.  The
function will then undefine itself.

This is especially useful for dynamically generated actions needed as a
result of an enter script, instead of writing them into the leave script
itself.