history helper
This helper is designed to allow you to have shell history which is
localized to the directory where you invoke it.  That history will be
used for the directory and any of its subdirectories, unless they
themselves override this with their own configuration.

Basic usage looks like this:

    smartcd helper run history also-read ~/.bash_history
    smartcd helper run history localize __PATH__/.history

The argument passed to 'localize' is the file in which your history will
be stored.

The 'also-read' action will read in any files you list prior to setting
up the new history.  This allows you to have access to any number of
other history files, but to put them "further back" in your buffer.