'smartcd on-prompt' command
This is similar to setting PROMPT_COMMAND or precmd, but with a smartcd
twist.  The argument passed to this command is evaulated at each prompt,
so long as you remain in the current directory or any of its children.

The commands are stored independently of any existing prompt hooks,
including smartcd's own.  In addition to this, they are fully nestable.
That is, you can add on-prompt commands in subdirectories after you have
already configured some and they will _all_ run until you return to a
previous directory.

For example:

    ~/foo$ smartcd on-prompt 'echo test 1'
    test 1
    ~/foo$ cd bar
    test 1
    ~/foo/bar$ smartcd on-prompt 'echo test 2'
    test 1
    test 2
    ~/foo/bar$ cd ..
    test 1
    ~/foo$ cd
    ~$