arrays file needed for _smartcd_file using zsh
When installing, the config command will call _smartcd_file() to
load the smartcd_config file. If a user is running zsh the anth()
function is used but the arrays file it comes from isn't yet
loaded. I've noticed that arrays is source in the config so it may
be worth while just sourceing it inside lib/core/smartcd but I
don't know the code well enough so I opted for this non-technical
fix.